refactor: rename schemas module
